cleaning/requirements.txt
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.2
SQLAlchemy==1.4.46
numpy==1.24.1
pandas==1.5.2
shapely==2.0.0